Update
MAKE_KEYBUFF is no longer descriptive.  We should choose a new name.  I
propose one here, but it could be better.

RT# are harder to follow then RT$.  I'm not sure it is worth adding
complexity to increase readability of the expanded program.  Completely
up for discussion.

To figure out if if a new record is possible, use MAX Recs and current
record count.  It's more robust and works for keyed records as well.